+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -314,7 +314,7 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             navigationBarBackgroundSchemeColor:
                 controller.navBarBackgroundSchemeColor,
             navigationBarOpacity: controller.navBarOpacity,
-            navigationBarElevation: controller.navigationBarElevation,
+            navigationBarElevation: controller.navBarElevation,
             navigationBarHeight: controller.navBarHeight,
             navigationBarLabelBehavior: controller.navBarLabelBehavior,
             // NavigationRail settings.