+++ example/lib/example5_themes_playground/widgets/panels/premade_designs/premade_designs_panel.dart
@@ -28,8 +28,7 @@ class PremadeDesignsPanel extends StatelessWidget {
           title: Text('Premade FlexColorScheme designs'),
           subtitle: Text('To quickly get started with theming, try these '
               'premade designs as starting points. You can use them as '
-              'inspiration and to see what FlexColorScheme can do.'
-              '\n\n'
+              'inspiration and to see what FlexColorScheme can do. '
               'These templates do no set colors, you choose them as desired. '
               'Most choices use seeded M3 ColorScheme and Material 3 mode. '
               'You can easily turn both off. In fact all configs are just '
@@ -37,7 +36,7 @@ class PremadeDesignsPanel extends StatelessWidget {
         ),
         const Divider(),
         SetupListTile(
-          title: 'Default settings',
+          title: 'Playground default',
           subtitle: 'Set settings back to their Themes Playground defaults. '
               'In Material 2 mode the default setup is quite opinionated, it '
               'mimics M3 in M2 mode. The M3 mode is almost default M3 style, '
@@ -48,7 +47,7 @@ class PremadeDesignsPanel extends StatelessWidget {
         ),
         const Divider(),
         SetupListTile(
-          title: 'Material 3 default styles',
+          title: 'Material 3 default',
           subtitle: 'Set Playground settings to a configurations that '
               'matches Material 3 defaults. Seeded color scheme is '
               'activated using the default Flutter Material 3 seeded '
@@ -58,37 +57,86 @@ class PremadeDesignsPanel extends StatelessWidget {
         ),
         const Divider(),
         SetupListTile(
-          title: 'Theme with light surface blends',
-          subtitle: 'This...',
+          title: 'Primary focused',
+          subtitle: 'This theme features primary color on navigators, filled '
+              'text input with primary opacity, rounders corner and no '
+              'unfocused border. '
+              'Outlines on buttons with outline are also primary colored. '
+              'Uses light surface blends with tinted texts and interactions. '
+              'It does not use seed generated ColorScheme',
           settingsId: 2,
           controller: controller,
         ),
         const Divider(),
         SetupListTile(
-          title: 'Theme with heavy background blends',
-          subtitle: 'This...',
+          title: 'Fabulous 12',
+          subtitle: 'Primary leaning theme, but with high scaffold blend, '
+              'a ColorScheme using 3 seed colors and the saturated Vivid '
+              'seed strategy. With smooth looking border radius 12 on all '
+              'widgets, except FAB, that is back in shape with its M2 circle.',
           settingsId: 3,
           controller: controller,
         ),
         const Divider(),
         SetupListTile(
-          title: 'Material 2 inspired Material 3',
-          subtitle: 'This...',
+          title: 'Material 3 does Material 2',
+          subtitle: 'Shows how you can emulate Material 2 style '
+              'in Material 3 mode. It does not use surface blends or seeded '
+              'ColorScheme. It removes elevation tint in light mode, but '
+              'keeps it in dark where it is useful. It brings back elevation '
+              'shadows in both light and dark mode if they were removed in M3. '
+              'It is not entirely default M2, it keeps some useful '
+              'consistency cleanups for a bit more refined M2 than Flutter '
+              'defaults. Need an M2 style M3 to migrate? Here it is!',
           settingsId: 4,
           controller: controller,
         ),
         const Divider(),
         SetupListTile(
-          title: 'Material 3 with bright colors',
-          subtitle: 'This...',
+          title: 'High contrast, locked keys',
+          subtitle: 'A theme that looks like it is not seed generated, but '
+              'it is, using the high contrast config and the three key colors '
+              'locked in light theme mode, also uses black and white on '
+              'colors in light mode. It has high surface blends, but since '
+              'true black and plain white setting is used, it is not very '
+              'visible and we have ink black scaffold in dart mode and '
+              'plain white one in light mode.',
           settingsId: 5,
           controller: controller,
         ),
         const Divider(),
         SetupListTile(
-          title: 'More platform agnostic Material 3',
-          subtitle: 'This...',
+          title: 'All primary based',
+          subtitle: 'This theme limits input colors to only the primary color, '
+              'It uses the One Hue seed generation to create a bright '
+              'ColorScheme where all colors are shades of the same primary '
+              'input. Even if you turn off using seeds, FlexColorScheme '
+              'computes all colors based on the primary input color.',
           settingsId: 6,
+          controller: controller,
+        ),
+        const Divider(),
+        SetupListTile(
+          title: 'Platform agnostic M3',
+          subtitle: 'This theme show a more platform agnostic design with M3, '
+              'it is less rounded then M3 default to feel more at home on '
+              'iOS but also it on Android ads brand expression. It removes '
+              'elevation tint in light mode, and puts shadows back in light '
+              'and dark mode, but this only happens on macOS and iOS, other '
+              'platforms keep the M3 surface tint.',
+          settingsId: 7,
+          controller: controller,
+        ),
+        const Divider(),
+        SetupListTile(
+          title: 'Computed dark theme',
+          subtitle: 'This theme is not seeded, but uses only three input '
+              'colors primary, secondary and tertiary from light mode and '
+              'computes all other colors, including all dark mode colors. If '
+              "you don't want to use seeded color and only want to specify "
+              'the main light theme colors, maybe 3 brand colors, and create '
+              'a full M3 ColorScheme based on them, this is an optional way.',
+          settingsId: 7,
           controller: controller,
         ),
         const Divider(),
@@ -135,13 +183,22 @@ class SetupListTile extends StatelessWidget {
   final int settingsId;
   final ThemeController controller;
 
-  Future<void> _handleSetToPreMade(BuildContext context, String title) async {
-    final bool? setToPremadeDesign = await showDialog<bool?>(
-      context: context,
-      builder: (BuildContext context) {
-        return SetThemeToPremadeDialog(selectedOption: title);
-      },
-    );
+  Future<void> _handleSetToPreMade(
+    BuildContext context,
+    String title, [
+    bool confirm = true,
+  ]) async {
+    bool? setToPremadeDesign;
+    if (confirm) {
+      setToPremadeDesign = await showDialog<bool?>(
+        context: context,
+        builder: (BuildContext context) {
+          return SetThemeToPremadeDialog(selectedOption: title);
+        },
+      );
+    } else {
+      setToPremadeDesign = true;
+    }
     if (setToPremadeDesign ?? false) {
       await controller.setToPremade(settingsId: settingsId);
     }
@@ -154,12 +211,12 @@ class SetupListTile extends StatelessWidget {
       subtitle: Text(subtitle),
       trailing: FilledButton(
         onPressed: () async {
-          await _handleSetToPreMade(context, title);
+          await _handleSetToPreMade(context, title, false);
         },
         child: const Text('Use'),
       ),
       onTap: () async {
-        await _handleSetToPreMade(context, title);
+        await _handleSetToPreMade(context, title, false);
       },
     );
   }
@@ -173,7 +230,7 @@ class SetThemeToPremadeDialog extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AlertDialog(
-      title: Text('Use ${selectedOption.toLowerCase()}?'),
+      title: Text('Use "$selectedOption" theme?'),
       content: const Text('Your current configuration will be overwritten, '
           'but custom colors will be kept.'),
       actions: <Widget>[
@@ -186,7 +243,7 @@ class SetThemeToPremadeDialog extends StatelessWidget {
             onPressed: () {
               Navigator.of(context).pop(true);
             },
-            child: const Text('Use design')),
+            child: const Text('Use theme')),
       ],
     );
   }